+++ backend/mocks/__init__.py
@@ -1,0 +1,3 @@
+# /mock/__init__.py
+
+from .mock_data import fake_database
