+++ backend/mocks/mock_data.py
@@ -1,0 +1,33 @@
+from models.models import PDFFile
+
+fake_database = [
+    PDFFile(id=1, filename="Student_1_Answers.pdf", filepath="Student_1_Answers.pdf"),
+    PDFFile(id=2, filename="Student_2_Answers.pdf", filepath="Student_2_Answers.pdf"),
+    PDFFile(id=3, filename="Student_3_Answers.pdf", filepath="Student_3_Answers.pdf"),
+    PDFFile(id=4, filename="Student_4_Answers.pdf", filepath="Student_4_Answers.pdf"),
+    PDFFile(id=5, filename="Student_5_Answers.pdf", filepath="Student_5_Answers.pdf"),
+    PDFFile(id=6, filename="Student_6_Answers.pdf", filepath="Student_6_Answers.pdf"),
+    PDFFile(id=7, filename="Student_7_Answers.pdf", filepath="Student_7_Answers.pdf"),
+    PDFFile(id=8, filename="Student_8_Answers.pdf", filepath="Student_8_Answers.pdf"),
+    PDFFile(id=9, filename="Student_9_Answers.pdf", filepath="Student_9_Answers.pdf"),
+    PDFFile(id=10, filename="Student_10_Answers.pdf", filepath="Student_10_Answers.pdf"),
+    PDFFile(id=11, filename="Student_11_Answers.pdf", filepath="Student_11_Answers.pdf"),
+    PDFFile(id=12, filename="Student_12_Answers.pdf", filepath="Student_12_Answers.pdf"),
+    PDFFile(id=13, filename="Student_13_Answers.pdf", filepath="Student_13_Answers.pdf"),
+    PDFFile(id=14, filename="Student_14_Answers.pdf", filepath="Student_14_Answers.pdf"),
+    PDFFile(id=15, filename="Student_15_Answers.pdf", filepath="Student_15_Answers.pdf"),
+    PDFFile(id=16, filename="Student_16_Answers.pdf", filepath="Student_16_Answers.pdf"),
+    PDFFile(id=17, filename="Student_17_Answers.pdf", filepath="Student_17_Answers.pdf"),
+    PDFFile(id=18, filename="Student_18_Answers.pdf", filepath="Student_18_Answers.pdf"),
+    PDFFile(id=19, filename="Student_19_Answers.pdf", filepath="Student_19_Answers.pdf"),
+    PDFFile(id=20, filename="Student_20_Answers.pdf", filepath="Student_20_Answers.pdf"),
+    PDFFile(id=21, filename="Student_21_Answers.pdf", filepath="Student_21_Answers.pdf"),
+    PDFFile(id=22, filename="Student_22_Answers.pdf", filepath="Student_22_Answers.pdf"),
+    PDFFile(id=23, filename="Student_23_Answers.pdf", filepath="Student_23_Answers.pdf"),
+    PDFFile(id=24, filename="Student_24_Answers.pdf", filepath="Student_24_Answers.pdf"),
+    PDFFile(id=25, filename="Student_25_Answers.pdf", filepath="Student_25_Answers.pdf"),
+    PDFFile(id=26, filename="Student_26_Answers.pdf", filepath="Student_26_Answers.pdf"),
+    PDFFile(id=27, filename="Student_27_Answers.pdf", filepath="Student_27_Answers.pdf"),
+    PDFFile(id=28, filename="Student_28_Answers.pdf", filepath="Student_28_Answers.pdf"),
+    PDFFile(id=29, filename="Student_29_Answers.pdf", filepath="Student_29_Answers.pdf"),
+]
