+++ backend/models/models.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+class Phrase(BaseModel):
+    text: str
+    explanation: str
+
+class PDFFile(BaseModel):
+    id: int
+    filename: str
+    filepath: str
